separate flux linkage and inductance analyzers
