--- docs/source/EM_analyzers/init_SynR_inductance_analyzer.csv
@@ -1,2 +1,0 @@
-Arguments,Description
-configuration,"object of type SynR_Inductance_Config describing time step, JMAG configurations, etc."
