+++ docs/source/EM_analyzers/input_SynR_inductance_analyzer.csv
@@ -1,0 +1,5 @@
+Arguments,Description,Units
+I_hat,peak current,A
+path,study path name,
+study_name,study name,
+time_step,time step,s
