+++ docs/source/EM_analyzers/output_SynR_flux_linkage_analyzer.csv
@@ -1,0 +1,5 @@
+Output,Description
+results_folder,folder where csv outputs are stored
+study_name,name of study
+I_hat,peak current in A
+time_step,time step in s
