+++ docs/source/EM_analyzers/results_SynR_flux_linkage_analyzer.csv
@@ -1,0 +1,5 @@
+Result,Value,Unit
+results_folder,/eMach/examples/mach_eval_examples/SynR_eval/run_data/jmag_csv/,
+study_name,Example_SynR_Machine_Ind_SynR,
+I_hat,20,A
+time_step,0.0001,s
