+++ examples/mach_eval_examples/SynR_eval/SynR_evaluator.py
@@ -6,13 +6,14 @@ os.chdir(os.path.dirname(__file__))
 sys.path.append("../../../")
 
 from mach_eval import (MachineEvaluator, MachineDesign)
+from SynR_flux_linkage_step import SynR_flux_linkage_step
 from SynR_inductance_step import SynR_inductance_step
-from electromagnetic_step import electromagnetic_step
 from example_SynR_machine import Example_SynR_Machine, Machine_Op_Pt
 
 ############################ Create Evaluator #####################
 SynR_evaluator = MachineEvaluator(
     [
+        SynR_flux_linkage_step,
         SynR_inductance_step
     ]
 )
